[Crypt] Moving Crypt classes to a new Manager namespace for better organization.
