+++ src/Valkyrja/Application/Env.php
@@ -35,8 +35,8 @@ use Valkyrja\Cli\Middleware\Contract\CommandNotMatchedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\ExitedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\InputReceivedMiddleware;
 use Valkyrja\Cli\Middleware\Contract\ThrowableCaughtMiddleware;
-use Valkyrja\Crypt\Contract\Crypt;
-use Valkyrja\Crypt\SodiumCrypt;
+use Valkyrja\Crypt\Manager\Contract\Crypt;
+use Valkyrja\Crypt\Manager\SodiumCrypt;
 use Valkyrja\Filesystem\Manager\Contract\Filesystem;
 use Valkyrja\Filesystem\Manager\FlysystemFilesystem;
 use Valkyrja\Filesystem\Manager\LocalFlysystemFilesystem;
