+++ tests/Unit/Auth/Provider/ServiceProviderTest.php
@@ -27,7 +27,7 @@ use Valkyrja\Auth\Store\InMemoryStore;
 use Valkyrja\Auth\Store\NullStore;
 use Valkyrja\Auth\Store\OrmStore;
 use Valkyrja\Auth\TokenAuthenticator;
-use Valkyrja\Crypt\Contract\Crypt;
+use Valkyrja\Crypt\Manager\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Jwt\Manager\Contract\Jwt;
 use Valkyrja\Orm\Manager\Contract\Manager;
