+++ tests/Unit/Broadcast/Provider/ServiceProviderTest.php
@@ -22,7 +22,7 @@ use Valkyrja\Broadcast\LogBroadcaster;
 use Valkyrja\Broadcast\NullBroadcaster;
 use Valkyrja\Broadcast\Provider\ServiceProvider;
 use Valkyrja\Broadcast\PusherBroadcaster;
-use Valkyrja\Crypt\Contract\Crypt;
+use Valkyrja\Crypt\Manager\Contract\Crypt;
 use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
